backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
librosa
Pillow
requests
scikit-learn
